requirements.txt
aiofiles==23.2.1
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
anyio==4.8.0
astor==0.8.1
audioop-lts==0.2.1
azure-ai-translation-text==1.0.1
azure-core==1.32.0
backoff==2.2.1
beautifulsoup4==4.13.4
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.1
cryptography==44.0.2
cutword==0.1.1
cycler==0.12.1
decorator==5.2.1
deepl==1.18.0
defusedxml==0.7.1
distro==1.9.0
doclayout_yolo==0.0.3
dukpy==0.5.0
et_xmlfile==2.0.0
fastapi==0.115.11
ffmpy==0.5.0
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.0
gradio==5.20.1
gradio_client==1.7.2
greenlet==3.2.2
groovy==0.1.2
grpcio==1.73.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
isodate==0.7.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.9.0
kiwisolver==1.4.8
magika==0.6.2
markdown-it-py==3.0.0
markdownify==1.1.0
# Editable install with no version control (markitdown==0.1.2)
-e /Users/manyfish/Downloads/markitdown/packages/markitdown
MarkupSafe==2.1.5
matplotlib==3.10.1
mdurl==0.1.2
mem0ai==0.1.114
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
ollama==0.4.7
onnxruntime==1.22.0
openai==1.66.3
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.10.15
packaging==24.2
paddlepaddle==3.0.0
pandas==2.2.3
pdf2zh==1.7.9
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.1.0
playwright==1.52.0
portalocker==2.10.1
posthog==6.1.1
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyahocorasick==2.1.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.1
PyMuPDF==1.25.3
pyparsing==3.2.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
qdrant-client==1.14.3
requests==2.32.3
rich==13.9.4
ruff==0.9.10
safehttpx==0.1.6
scipy==1.15.2
seaborn==0.13.2
semantic-version==2.10.0
setuptools==76.0.0
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.1
stringzilla==3.12.3
sympy==1.13.1
tenacity==9.0.0
thop==0.1.1.post2209072238
tomlkit==0.13.2
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0.1
yt-dlp==2025.6.9
